database/requirements.txt
# Python dependencies for database setup scripts
# Install with: pip install -r requirements.txt

requests>=2.31.0
python-dotenv>=1.0.0
